move type definitions closer to consumer
