+++ apps/flame-defi/app/features/cosmos-wallet/hooks/use-ibc-currency-balance.ts
@@ -1,16 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
 import { Decimal } from "@cosmjs/math";
 
-import {
-  Balance,
-  IbcCurrency,
-  PollingConfig,
-  UseBalanceResult,
-} from "@repo/flame-types";
+import { Balance, IbcCurrency } from "@repo/flame-types";
 import { useConfig } from "config";
+import { PollingConfig, UseBalanceResult } from "hooks/use-currency-balance";
 
-import { getBalanceFromChain } from "../services/cosmos";
-import { useCosmosWallet } from "./use-cosmos-wallet";
+import { useCosmosWallet } from "features/cosmos-wallet/hooks/use-cosmos-wallet";
+import { getBalanceFromChain } from "features/cosmos-wallet/services/cosmos";
 
 /**
  * Custom hook to fetch and manage the balance for an IBC currency.
