+++ apps/flame-defi/app/features/evm-wallet/contexts/evm-wallet-context.tsx
@@ -1,4 +1,4 @@
-import { useConnectModal } from "@rainbow-me/rainbowkit";
+import { useChainModal, useConnectModal } from "@rainbow-me/rainbowkit";
 import React, { useCallback, useEffect, useState } from "react";
 import { useAccount, useDisconnect, useSwitchChain } from "wagmi";
 
@@ -22,6 +22,7 @@ export const EvmWalletProvider: React.FC<EvmWalletProviderProps> = ({
   children,
 }) => {
   const { openConnectModal } = useConnectModal();
+  const { openChainModal } = useChainModal();
   const { disconnect } = useDisconnect();
   const userAccount = useAccount();
   const { switchChain } = useSwitchChain();
@@ -48,9 +49,11 @@ export const EvmWalletProvider: React.FC<EvmWalletProviderProps> = ({
   );
 
   const connectToSpecificChain = (chainId: number) => {
-    if (!evmAccountAddress && openConnectModal) {
-      openConnectModal();
-    } else if (!evmAccountAddress) {
+    console.log({ evmAccountAddress, chainId });
+    if (!evmAccountAddress) {
+      // FIXME - how do i switch chains after they connect?
+      openConnectModal?.();
+    } else {
       switchChain({ chainId });
     }
   };
